pyproject.toml
[tool.poetry]
name = "csv-to-db"
version = "0.1.0"
description = ""
authors = ["Marc Ford <mrfxyz567@gmail.com>"]
license = "MIT"
readme = "README.md"
repository = "https://github.com/mfdeux/csv-to-db"
packages = [{ include = "csv_to_db" }]
include = ["data/*"]

[tool.poetry.dependencies]
python = "^3.8"
requests = "^2.24.0"
click = "^7.1.2"
tqdm = "^4.48.2"
GitPython = "^3.1.7"
sqlalchemy = "^1.3.18"
watchdog = "^0.10.3"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
ipython = "^7.17.0"
black = "^19.10b0"

[tool.poetry.scripts]
csv-to-db = "csv_to_db.cli:cli"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
